requirements.txt
fastapi
uvicorn
pydantic
email-validator
google-generativeai
